Rename partialResponse middleware by httpPartialResponse
