+++ src/middlewares/index.js
@@ -6,8 +6,8 @@ module.exports = {
   httpErrorHandler: require('./httpErrorHandler'),
   httpEventNormalizer: require('./httpEventNormalizer'),
   httpHeaderNormalizer: require('./httpHeaderNormalizer'),
+  httpPartialResponse: require('./httpPartialResponse'),
   jsonBodyParser: require('./jsonBodyParser'),
-  partialResponse: require('./partialResponse'),
   s3KeyNormalizer: require('./s3KeyNormalizer'),
   ssm: require('./ssm'),
   urlEncodeBodyParser: require('./urlEncodeBodyParser'),
